+++ src/AppForSEII2526.API/ClassDiagram.cd
@@ -1,0 +1,1 @@
+<?xml version="1.0" encoding="utf-8"?><ClassDiagram></ClassDiagram>
