--- src/AppForSEII2526.API/ClassDiagramCDU1.cd
@@ -1,58 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ClassDiagram MajorVersion="1" MinorVersion="1">
-  <Class Name="AppForSEII2526.API.Models.fabricante">
-    <Position X="1" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAACAAAAAAAAAAAAgAAAAAAAAAAAAAgA=</HashCode>
-      <FileName>Models\fabricante.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsCollectionAssociation>
-      <Property Name="herramientas" />
-    </ShowAsCollectionAssociation>
-  </Class>
-  <Class Name="AppForSEII2526.API.Models.Herramienta">
-    <Position X="4" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAACAAAAAAAAAEAAAMAAAEAAAAIBAEgA=</HashCode>
-      <FileName>Models\Herramienta.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="fabricante" />
-    </ShowAsAssociation>
-    <ShowAsCollectionAssociation>
-      <Property Name="ComprarItems" />
-    </ShowAsCollectionAssociation>
-  </Class>
-  <Class Name="AppForSEII2526.API.Models.ComprarItem">
-    <Position X="7.75" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AQAAAAAAAAAAAAAAAAEAEAAAAAAgEAQAAAAAAAAABAA=</HashCode>
-      <FileName>Models\ComprarItem.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="herramienta" />
-      <Property Name="compra" />
-    </ShowAsAssociation>
-  </Class>
-  <Class Name="AppForSEII2526.API.Models.Compra">
-    <Position X="11" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>ABACAAAAAAgAABAAAAIAAAAACABAAAAAAAAQAAAAAgI=</HashCode>
-      <FileName>Models\Compra.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="metodopago" />
-    </ShowAsAssociation>
-    <ShowAsCollectionAssociation>
-      <Property Name="CompraItems" />
-    </ShowAsCollectionAssociation>
-  </Class>
-  <Enum Name="AppForSEII2526.API.Models.formaPago">
-    <Position X="10.25" Y="4.25" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAABAAAAAAAAAAAAAAAgAAAAAAAABAAA=</HashCode>
-      <FileName>Models\Compra.cs</FileName>
-    </TypeIdentifier>
-  </Enum>
-  <Font Name="Segoe UI" Size="9" />
-</ClassDiagram>
