--- src/AppForSEII2526.API/ClassDiagramCDU2.cd
@@ -1,58 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ClassDiagram MajorVersion="1" MinorVersion="1">
-  <Class Name="AppForSEII2526.API.Models.fabricante">
-    <Position X="0.5" Y="0.75" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAACAAAAAAAAAAAAgAAAAAAAAAAAAAgA=</HashCode>
-      <FileName>Models\fabricante.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsCollectionAssociation>
-      <Property Name="herramientas" />
-    </ShowAsCollectionAssociation>
-  </Class>
-  <Class Name="AppForSEII2526.API.Models.Herramienta">
-    <Position X="3.25" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAACAAAAAAAAAEAAAMAAAEAAAAIBAEgA=</HashCode>
-      <FileName>Models\Herramienta.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="fabricante" />
-    </ShowAsAssociation>
-    <ShowAsCollectionAssociation>
-      <Property Name="ReparacionItem" />
-    </ShowAsCollectionAssociation>
-  </Class>
-  <Class Name="AppForSEII2526.API.Models.Reparacion">
-    <Position X="9.25" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>ABACAAAAAIgAAAAAAAAAAAEAAAAAAAAEAAAQAIAAAgA=</HashCode>
-      <FileName>Models\Reparacion.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="metodoPago" />
-    </ShowAsAssociation>
-    <ShowAsCollectionAssociation>
-      <Property Name="ReparacionItems" />
-    </ShowAsCollectionAssociation>
-  </Class>
-  <Class Name="AppForSEII2526.API.Models.ReparacionItem">
-    <Position X="6.25" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AQAAAAAAAAAAAAAAAAEAEAAAAAAAEAgAIAAAAAAAAEA=</HashCode>
-      <FileName>Models\ReparacionItem.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="Herramienta" />
-      <Property Name="Reparacion" />
-    </ShowAsAssociation>
-  </Class>
-  <Enum Name="AppForSEII2526.API.Models.metodoPago">
-    <Position X="7.5" Y="3" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAABAAAAAAAAAAAAAAAgAAAAAAAABAAA=</HashCode>
-      <FileName>Models\Reparacion.cs</FileName>
-    </TypeIdentifier>
-  </Enum>
-  <Font Name="Segoe UI" Size="9" />
-</ClassDiagram>
