--- src/AppForSEII2526.API/ClassDiagramCDU3.cd
@@ -1,71 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ClassDiagram MajorVersion="1" MinorVersion="1">
-  <Class Name="AppForSEII2526.API.Models.fabricante">
-    <Position X="0.5" Y="1.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAACAAAAAAAAAAAAgAAAAAAAAAAAAAgA=</HashCode>
-      <FileName>Models\fabricante.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsCollectionAssociation>
-      <Property Name="herramientas" />
-    </ShowAsCollectionAssociation>
-  </Class>
-  <Class Name="AppForSEII2526.API.Models.Herramienta">
-    <Position X="4" Y="1.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAACAAAAAAAAAEAAAMAAAEAAAAIBAEgA=</HashCode>
-      <FileName>Models\Herramienta.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="fabricante" />
-    </ShowAsAssociation>
-    <ShowAsCollectionAssociation>
-      <Property Name="OfertaItems" />
-    </ShowAsCollectionAssociation>
-  </Class>
-  <Class Name="AppForSEII2526.API.Models.OfertaItem">
-    <Position X="8" Y="1.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAAAAAAAEAAAAAAgIAABAAgAAAAAgAA=</HashCode>
-      <FileName>Models\OfertaItem.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="herramienta" />
-      <Property Name="oferta" />
-    </ShowAsAssociation>
-  </Class>
-  <Class Name="AppForSEII2526.API.Models.Oferta">
-    <Position X="11.5" Y="1.5" Width="1.5" />
-    <AssociationLine Name="metodoPago" Type="AppForSEII2526.API.Models.tiposMetodoPago">
-      <MemberNameLabel ManuallyPlaced="true">
-        <Position X="0.006" Y="0.081" />
-      </MemberNameLabel>
-    </AssociationLine>
-    <TypeIdentifier>
-      <HashCode>AAACAAAAAAAAIAAAAAACAAAAAAAAQBAAACAAAAAAAAA=</HashCode>
-      <FileName>Models\Oferta.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="metodoPago" />
-      <Property Name="paraSocio" />
-    </ShowAsAssociation>
-    <ShowAsCollectionAssociation>
-      <Property Name="ofertaItems" />
-    </ShowAsCollectionAssociation>
-  </Class>
-  <Enum Name="AppForSEII2526.API.Models.tiposMetodoPago">
-    <Position X="13.5" Y="4.25" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAACAABAAAAAAAAAAAAAAAgAAAAAAAAAAAA=</HashCode>
-      <FileName>Models\Oferta.cs</FileName>
-    </TypeIdentifier>
-  </Enum>
-  <Enum Name="AppForSEII2526.API.Models.tiposDirigidaOferta">
-    <Position X="9.25" Y="4.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAABAAAAAAAA=</HashCode>
-      <FileName>Models\Oferta.cs</FileName>
-    </TypeIdentifier>
-  </Enum>
-  <Font Name="Segoe UI" Size="9" />
-</ClassDiagram>
