+++ src/AppForSEII2526.API/Models/OfertaItem.cs
@@ -1,23 +1,16 @@
 ﻿namespace AppForSEII2526.API.Models
 {
-
-    [PrimaryKey(nameof(idOferta), nameof(idHerramienta))]
     public class OfertaItem
     {
-        
+        [Key]
         public int idOferta { get; set; }
 
         public int idHerramienta { get; set; }
 
-        [Required, Range(1, 100, ErrorMessage = "Establece un porcentaje entre 1 y 100")]
+        [Range(1, 100, ErrorMessage = "Establece un porcentaje entre 1 y 100")]
         public int porcentaje { get; set; }
 
         [DataType(System.ComponentModel.DataAnnotations.DataType.Currency)]
         public float precioFinal { get; set; }
-
-        // Relaciones
-        public Oferta oferta { get; set; }
-        public Herramienta herramienta { get; set; }
-
     }
 }
