+++ src/AppForSEII2526.API/Models/fabricante.cs
@@ -6,10 +6,7 @@
         public int id { get; set; }
 
 
-        [StringLength(50, ErrorMessage = "No puede tener mas de 50 caracteres.", MinimumLength = 1)]
+        [StringLength(50, ErrorMessage = "No puede tener mas de 50 caracteres.", MinimumLength = 4)]
         public string nombre { get; set; }
-
-        public IList<Herramienta> herramientas { get; set; }
-
     }
 }
